add mixin generator
In order to give specific permissions to a group or user, an appropriate
table mapping user/group to the resource along with the permissions is
needed. This module has a function which generates a mixin for creating
that table. It follows the SQLAlchemy Association Object
(https://docs.sqlalchemy.org/en/13/orm/basic_relationships.html#association-object)
pattern